+++ lib/workmanager.dart
@@ -42,8 +42,8 @@ class WorkManagerConstraintConfig {
 
 const _noDuration = const Duration(seconds: 0);
 
-/// Returns the echo value provided when registering the task.
-typedef EchoCallbackFunction = Future<bool> Function(String echoValue);
+/// Returns the value you provided when registering the task. iOS will always return [Workmanager.iOSBackgroundTask]
+typedef BackgroundTask = Future<bool> Function(String echoValue);
 
 ///An enumeration of the conflict resolution policies in case of a collision.
 enum ExistingWorkPolicy {
@@ -86,28 +86,61 @@ enum BackoffPolicy {
   linear
 }
 
+/// Workmanager plugin entrypoint
+///
+/// Make sure you followed the setup steps first before trying to register any task.
+///
+/// Initialize the plugin first
+///
+/// ```
+/// void callbackDispatcher() {
+///   Workmanager.executeTask((echoValue) {
+///     print("Native echoed: $echoValue");
+///     return Future.value(true);
+///   });
+/// }
+///
+/// Workmanager.initialize(
+///     callbackDispatcher,
+///     isInDebugMode: true
+/// )
+/// ```
 class Workmanager {
+  /// Use this constant inside your callbackDispatcher to identify when an iOS Background Fetch occurred.
+  ///
+  /// ```
+  /// void callbackDispatcher() {
+  ///  Workmanager.executeTask((task) async {
+  ///      switch (task) {
+  ///        case Workmanager.iOSBackgroundTask:
+  ///          stderr.writeln("The iOS background fetch was triggered");
+  ///          break;
+  ///      }
+  ///
+  ///      return Future.value(true);
+  ///  });
+  /// }
+  /// ```
+  static const String iOSBackgroundTask = "iOSPerformFetch";
   static bool _isInDebugMode = false;
 
   static const MethodChannel _backgroundChannel =
-      const MethodChannel(_WorkmanagerConstants.backgroundChannelName);
+  const MethodChannel(_WorkmanagerConstants.backgroundChannelName);
   static const MethodChannel _foregroundChannel =
-      const MethodChannel(_WorkmanagerConstants.foregroundChannelName);
+  const MethodChannel(_WorkmanagerConstants.foregroundChannelName);
 
-  /// A helper function so you only need to implement a [EchoCallbackFunction]
-  static void defaultCallbackDispatcher(
-      final EchoCallbackFunction echoFunction) {
+  /// A helper function so you only need to implement a [BackgroundTask]
+  static void executeTask(final BackgroundTask backgroundTask) {
     WidgetsFlutterBinding.ensureInitialized();
     _backgroundChannel
-        .setMethodCallHandler((call) async => echoFunction(call.arguments));
+        .setMethodCallHandler((call) async => backgroundTask(call.arguments));
     _backgroundChannel.invokeMethod("backgroundChannelInitialized");
   }
 
-  /// This call is required if you wish to use the WorkManager plugin.
-  /// callbackDispatcher is a top level function which will be invoked by Android
-  /// isInDebugMode true will post debug notifications with information about when a job should have run
-  static Future<void> initialize(
-    final Function callbackDispatcher, {
+  /// This call is required if you wish to use the [WorkManager] plugin.
+  /// [callbackDispatcher] is a top level function which will be invoked by Android
+  /// [isInDebugMode] true will post debug notifications with information about when a task should have run
+  static Future<void> initialize(final Function callbackDispatcher, {
         final bool isInDebugMode = false,
   }) async {
     Workmanager._isInDebugMode = isInDebugMode;
@@ -116,22 +149,21 @@ class Workmanager {
   }
 
   /// Schedule a one off task
-  /// A unique name is required so only one job can be registered.
-  /// The echoValue is the value that will be returned in the [EchoCallbackFunction]
-  static Future<void> registerOneOffTask(
-    final String uniqueName,
-    final String echoValue, {
-    final String tag,
-    final ExistingWorkPolicy existingWorkPolicy,
-    final Duration initialDelay = _noDuration,
-    final WorkManagerConstraintConfig constraints,
-    final BackoffPolicy backoffPolicy,
-    final Duration backoffPolicyDelay = _noDuration,
-  }) async =>
+  /// A [uniqueName] is required so only one task can be registered.
+  /// The [valueToReturn] is the value that will be returned in the [BackgroundTask]
+  static Future<void> registerOneOffTask(final String uniqueName,
+      final String valueToReturn, {
+        final String tag,
+        final ExistingWorkPolicy existingWorkPolicy,
+        final Duration initialDelay = _noDuration,
+        final WorkManagerConstraintConfig constraints,
+        final BackoffPolicy backoffPolicy,
+        final Duration backoffPolicyDelay = _noDuration,
+      }) async =>
       await _register(
         methodName: "registerOneOffTask",
         uniqueName: uniqueName,
-        echoValue: echoValue,
+        echoValue: valueToReturn,
         tag: tag,
         existingWorkPolicy: existingWorkPolicy,
         initialDelay: initialDelay,
@@ -141,20 +173,19 @@ class Workmanager {
       );
 
   /// Schedules a periodic task that will run every provided [frequency].
-  /// A unique name is required so only one job can be registered.
-  /// The echoValue is the value that will be returned in the [EchoCallbackFunction]
-  /// a frequency is not required and will be defaulted to 15 minutes if not provided.
-  static Future<void> registerPeriodicTask(
-    final String uniqueName,
-    final String echoValue, {
-    final Duration frequency,
-    final String tag,
-    final ExistingWorkPolicy existingWorkPolicy,
-    final Duration initialDelay = _noDuration,
-    final WorkManagerConstraintConfig constraints,
-    final BackoffPolicy backoffPolicy,
-    final Duration backoffPolicyDelay = _noDuration,
-  }) async =>
+  /// A [uniqueName] is required so only one task can be registered.
+  /// The [echoValue] is the value that will be returned in the [BackgroundTask]
+  /// a [frequency] is not required and will be defaulted to 15 minutes if not provided.
+  static Future<void> registerPeriodicTask(final String uniqueName,
+      final String echoValue, {
+        final Duration frequency,
+        final String tag,
+        final ExistingWorkPolicy existingWorkPolicy,
+        final Duration initialDelay = _noDuration,
+        final WorkManagerConstraintConfig constraints,
+        final BackoffPolicy backoffPolicy,
+        final Duration backoffPolicyDelay = _noDuration,
+      }) async =>
       await _register(
         methodName: "registerPeriodicTask",
         uniqueName: uniqueName,
@@ -203,16 +234,16 @@ class Workmanager {
     );
   }
 
-  /// Cancels a job by its unique name
+  /// Cancels a task by its [uniqueName]
   static Future<void> cancelByUniqueName(final String uniqueName) async =>
       await _foregroundChannel
           .invokeMethod("cancelTaskByUniqueName", {"uniqueName": uniqueName});
 
-  /// Cancels a job by its tag
+  /// Cancels a task by its [tag]
   static Future<void> cancelByTag(final String tag) async =>
       await _foregroundChannel.invokeMethod("cancelTaskByTag", {"tag": tag});
 
-  /// Cancels all jobs
+  /// Cancels all tasks
   static Future<void> cancelAll() async =>
       await _foregroundChannel.invokeMethod("cancelAll");
 }